Extract Electrode/ElectrodeType to lightweight electrodes.py module
Can be imported without NEURON/MPI: from bluerecording.electrodes import Electrode
